feat: extract client type vars into types package
