+++ python/mirascope/llm/types/__init__.py
@@ -1,5 +1,6 @@
 """Types for the LLM module."""
 
+from .clients import LLMT, ClientT, ParamsT, ProviderMessageT
 from .dataclass import Dataclass
 from .jsonable import Jsonable
 from .prompts import PromptT
@@ -8,18 +9,20 @@ from .type_vars import (
     DepsT,
     FormatT,
     P,
-    ProviderMessageT,
     RequiredFormatT,
     ToolReturnT,
 )
 
 __all__ = [
+    "LLMT",
+    "ClientT",
     "CovariantT",
     "Dataclass",
     "DepsT",
     "FormatT",
     "Jsonable",
     "P",
+    "ParamsT",
     "PromptT",
     "ProviderMessageT",
     "RequiredFormatT",
